+++ app/src/main/java/com/reddit/rickandmortyapp/noncompose/NonComposeActivity.kt
@@ -4,9 +4,9 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.reddit.rickandmortyapp.R
 
-class MainActivity : AppCompatActivity() {
+class NonComposeActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_main_non_compose)
     }
 }
