+++ app/src/main/res/layout/activity_main_non_compose.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".noncompose.MainActivity">
+    tools:context=".noncompose.NonComposeActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="0dp"
